abstract shibboleth login for non KIT
